requirements.txt
google-genai>=1.0
openai>=1.30
anthropic>=0.30
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
httpx>=0.27
feedparser>=6.0
manim>=0.18
moviepy>=1.0
Pillow>=10.0
schedule>=1.2
jinja2>=3.1
click>=8.1
